Fix an issue where a truncate would not get propagated to userspace
In the event that an operation moved a file to overwrite an existing file (such as the following example)
```
echo "foo" > foo
echo "bar" > bar
mv foo bar
```
A delete event would be sent for `foo`, and a create event would be sent for the new `bar`, but no delete event would be sent for the original `bar` file that was overwritten. This change updates the behavior to send this delete event if the target destination exists.

Also fixes an issue where the inode reported in the create event of the new `bar` would actually report the inode of the old file. The correct inode to report is the one associated with `foo` before moving the file, since the inode of the source is constant across the rename.

Signed-off-by: Dom Bozzuto <dbozzuto@vmware.com>